requirements.txt
Flask
PyJWT
Flask-Mail
Flask-SQLAlchemy
Flask-Limiter
python-dotenv
